% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{C_get_W}
\alias{C_get_W}
\title{Calculate the W statistic for the Royston approach (implemented in Rcpp)}
\usage{
C_get_W(vec_value)
}
\arguments{
\item{vec_value}{vector containing data points; integer or double}
}
\value{
the test statistic for the Shapiro-Wilk test; double, between 0 and
1
}
\description{
This function calculates the W test statistic for the Royston approach for
the Shapiro-Wilk test. This function is comparable to the R implementation
of the Royston approach (which can be found in the R folder). For a
step-by-step explanation of this function, look at the Readme file under
"A Note on the Algorithms".
}
\examples{
C_get_W(1:10)

}
