% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_test.R
\name{check_sw_test_inputs}
\alias{check_sw_test_inputs}
\title{Check for valid parameters for a sw_test call}
\usage{
check_sw_test_inputs(vec_value, approach)
}
\arguments{
\item{vec_value}{vector}

\item{approach}{character}
}
\value{
approach_to_use
}
\description{
Check for valid parameters for a sw_test call
}
\examples{
check_sw_test_inputs (1:60, "royston")
}
