% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_test.R
\name{find_surrounding_pair}
\alias{find_surrounding_pair}
\title{Find the two values within a vector that immediately surround the target value}
\usage{
find_surrounding_pair(target, x)
}
\arguments{
\item{target}{the value (i.e. the area under the curve); double or integer}

\item{x}{a vector contain values (i.e. the row in the p-value chart
corresponding to n); double or integer corresponding to target}
}
\value{
a list with positions of the values within the vector and the values
in those positions
}
\description{
This function is used in the checknormality function to mimic how you use a
p-value chart.
}
\examples{
find_surrounding_pair (5, 1:10)
}
