% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pvalue.R
\name{get_pvalue}
\alias{get_pvalue}
\title{Calculate the p-value associated with the W statistic}
\usage{
get_pvalue(W, n)
}
\arguments{
\item{W}{double}

\item{n}{integer}
}
\value{
p-value
}
\description{
Note that this is specific for p-values using the original or modified original Shapiro-Wilk method
}
\examples{
get_pvalue (.970, 20)
}
