% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{NormalCDFInverse}
\alias{NormalCDFInverse}
\title{Calculates the Normal CDF Inverse}
\usage{
NormalCDFInverse(p)
}
\arguments{
\item{double}{p}
}
\value{
value for which the normal CDF < p
}
\description{
This function was taken from https://www.johndcook.com/blog/cpp_phi_inverse/
}
