% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RationalApproximation}
\alias{RationalApproximation}
\title{Multiplies two doubles}
\usage{
RationalApproximation(t)
}
\arguments{
\item{t}{double}
}
\value{
Product of v1 and v2
}
\description{
Multiplies two doubles
}
