% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_test.R
\name{royston_sw}
\alias{royston_sw}
\title{Implement the Royston approach for the Shapiro-Wilk test}
\usage{
royston_sw(vec_value)
}
\arguments{
\item{vec_value}{}
}
\value{
w_p list containing the W test-statistic and p-value
}
\description{
Implement the Royston approach for the Shapiro-Wilk test
}
\examples{
royston_sw (1:60)
}
