% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{stl_sort}
\alias{stl_sort}
\title{Sorts a vector in ascending order}
\usage{
stl_sort(x)
}
\arguments{
\item{NumericVector}{x}
}
\value{
x in ascending order
}
\description{
Sorts a vector in ascending order
}
