% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_test.R
\name{sw_test}
\alias{sw_test}
\title{Shapiro-Wilk Test}
\usage{
sw_test(x)
}
\arguments{
\item{x}{vector}
}
\value{
p_value
}
\description{
Test the normality of a vector using the original Shapiro-Wilk test
}
\examples{
sw_test (rnorm(100, 0, 1))

}
