% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_surrounding_pair.R
\name{find_surrounding_pair}
\alias{find_surrounding_pair}
\title{Find the two values within a vector that immediately surround the target value}
\usage{
find_surrounding_pair(target, x)
}
\arguments{
\item{target}{vector}

\item{x}{value (integer or double)}
}
\value{
a list with positions of the values within the vector and the values
}
\description{
Find the two values within a vector that immediately surround the target value
}
\examples{
find_surrounding_pair (5, 1:10)
}
