% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_test.R
\name{get_pvalue}
\alias{get_pvalue}
\title{Calculate the p-value associated with the W statistic}
\usage{
get_pvalue(W, n, use_harmonic = T)
}
\arguments{
\item{W}{the test statistic for the Shapiro-Wilk test; double, between 0 and
1}

\item{n}{the number of data points; integer}

\item{use_harmonic}{whether or not to use harmonic interpolation; boolean, 0
corresponding to linear interpolation and 1 corresponding to harmonic
interpolation}
}
\value{
p-value, double
}
\description{
This function uses interpolation to calculate the exact p-value using the
Shapiro-Wilk p-value table. Note that this is specific for p-values using the
 original or modified Shapiro-Wilk approach. P-values < 0.01 are reported as
 "0.001", and p-values > 0.99 are reported as "0.999".
}
\examples{
get_pvalue (.970, 20)
}
