% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pvalue.R
\name{get_pvalue}
\alias{get_pvalue}
\title{Calculate the p-value associated with the W statistic}
\usage{
get_pvalue(W, n, h = "harmonic")
}
\arguments{
\item{W}{double, the test statistic for the Shapiro-Wilk test}

\item{n}{integer, the number of data points}

\item{h}{character, whether or not to use harmonic interpolation; the alternative is linear interpolation}
}
\value{
p-value
}
\description{
Note that this is specific for p-values using the original or modified original Shapiro-Wilk method
Because this implementation uses a p-value table, p-values < 0.01 are reported as "0.001", and p-values > 0.99 are reported as "0.999"
}
\examples{
get_pvalue (.970, 20)
}
