% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pvalue.R
\name{get_pvalue}
\alias{get_pvalue}
\title{Calculate the p-value associated with the W statistic}
\usage{
get_pvalue(W)
}
\arguments{
\item{W}{double}
}
\value{
p-value
}
\description{
Calculate the p-value associated with the W statistic
}
\examples{
get_pvalue (.970)
}
