% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_test.R
\name{original_sw}
\alias{original_sw}
\title{Implement the original and modified approaches for the Shapiro-Wilk test}
\usage{
original_sw(vec_value, approach = "modified")
}
\arguments{
\item{vec_value}{double}

\item{approach}{character}
}
\value{
w_p list containing the W test-statistic and p-value
}
\description{
Implement the original and modified approaches for the Shapiro-Wilk test
}
\examples{
original_sw (1:60, "modified")
}
