% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_test.R
\name{R_get_W}
\alias{R_get_W}
\title{Get the W statistic for the Royston approach (implemented in R)}
\usage{
R_get_W(vec_value)
}
\arguments{
\item{vec_value}{}
}
\value{
W
}
\description{
Get the W statistic for the Royston approach (implemented in R)
}
\examples{
R_get_W (1:10)
}
