% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_test.R
\name{sw_test}
\alias{sw_test}
\title{Shapiro-Wilk Test}
\usage{
sw_test(vec_value, alpha = 0.05)
}
\arguments{
\item{vec_value}{vector}

\item{alpha}{character}
}
\value{
p_value
}
\description{
Test the normality of a vector using either the original or modified (Royston) version of the Shapiro-Wilk test
}
\examples{
sw_test (rnorm(10, 0, 1))

}
